Change int64 to uint64 for expiry block height

+++ x/profiles/keeper/keeper_app_links.go
@@ -129,13 +129,13 @@ func (k Keeper) DeleteAllUserApplicationLinks(ctx sdk.Context, user string) {
 
 // UpdateExpiringApplicationLinks updates the states of all the expiring application links to be expired
 func (k Keeper) UpdateExpiringApplicationLinks(ctx sdk.Context) {
-	k.IterateExpiringApplicationLinks(ctx, ctx.BlockHeight(), func(_ int64, link types.ApplicationLink) (stop bool) {
+	k.IterateExpiringApplicationLinks(ctx, uint64(ctx.BlockHeight()), func(_ int64, link types.ApplicationLink) (stop bool) {
 		store := ctx.KVStore(k.storeKey)
 		link.State = types.AppLinkStateVerificationExpired
 		userApplicationLinkKey := types.UserApplicationLinkKey(link.User, link.Data.Application, link.Data.Username)
 		store.Set(userApplicationLinkKey, types.MustMarshalApplicationLink(k.cdc, link))
 
-		store.Delete(types.ExpiringApplicationLinkKey(ctx.BlockHeight(), link.OracleRequest.ClientID))
+		store.Delete(types.ExpiringApplicationLinkKey(uint64(ctx.BlockHeight()), link.OracleRequest.ClientID))
 		return false
 	})
 }
